refactor(deepspeed): tcp for multi-node, file for single-node,add some notes on multi-node training

+++ examples/aishell/s0/run.sh
@@ -14,6 +14,7 @@ stop_stage=5
 # see https://pytorch.org/docs/stable/elastic/run.html
 HOST_NODE_ADDR="localhost:0"
 num_nodes=1
+job_id=2023
 
 # The aishell dataset location, please change this to your own path
 # make sure of using absolute path. DO-NOT-USE relatvie path!
@@ -114,11 +115,12 @@ fi
 
 if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
   mkdir -p $dir
-  # You have to rm `INIT_FILE` manually when you resume or restart a
-  # multi-machine training.
-  INIT_FILE=$dir/ddp_init
-  rm -f ${INIT_FILE}  # remove previous INIT_FILE
-  init_method=file://$(readlink -f $INIT_FILE)
+  init_method=tcp://$HOST_NODE_ADDR  # multi-node multi-gpu, tcp is recommended
+  if [ ${num_nodes} -le 1  ]; then
+    INIT_FILE=$dir/ddp_init
+    rm -f ${INIT_FILE}  # remove previous INIT_FILE
+    init_method=file://$(readlink -f $INIT_FILE)  # single-node single-gpu, file is recommended
+  fi
   echo "$0: init method is $init_method"
   num_gpus=$(echo $CUDA_VISIBLE_DEVICES | awk -F "," '{print NF}')
   # Use "nccl" if it works, otherwise use "gloo"
@@ -133,25 +135,29 @@ if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
   # and output dimension, and $dir/train.yaml will be used for inference
   # and export.
   if [ ${train_engine} == "deepspeed" ]; then
-    echo "using deepspeed"
-    # TODO(xcsong): remove filter and impl model.join for deepspeed
-    # [ ! -f data/$train_set/data.list.filter ] && \
-    #   python tools/filter_uneven_data.py data/$train_set/data.list \
-    #     $data_type $num_gpus $num_utts_per_shard data/$train_set/data.list.filter
+    echo "$0: using deepspeed"
   else
-    echo "using torch ddp"
+    echo "$0: using torch ddp"
   fi
 
   # NOTE(xcsong): Both ddp & deepspeed can be launched by torchrun
   # NOTE(xcsong): To unify single-node & multi-node training, we add
-  #               all `rdzv` related args. You should change `nnodes` &
+  #               all related args. You should change `nnodes` &
   #               `rdzv_endpoint` for multi-node, see
   #               https://pytorch.org/docs/stable/elastic/run.html#usage
   #               `rdzv_id` - A user-defined id that uniquely identifies the worker group for a job.
   #                           This id is used by each node to join as a member of a particular worker group.
   #               `rdzv_endpoint` - The rendezvous backend endpoint; usually in form <host>:<port>.
+  # NOTE(xcsong): In multi-node training, some clusters require special NCCL variables to set prior to training.
+  #               For example: `NCCL_IB_DISABLE=1` + `NCCL_SOCKET_IFNAME=enp` + `NCCL_DEBUG=INFO`
+  #               without NCCL_IB_DISABLE=1
+  #                   RuntimeError: NCCL error in: ../torch/csrc/distributed/c10d/ProcessGroupNCCL.cpp:1269, internal error, NCCL Version xxx
+  #               without NCCL_SOCKET_IFNAME=enp  (IFNAME could be get by `ifconfig`)
+  #                   RuntimeError: The server socket has failed to listen on any local network address. The server socket has failed to bind to [::]:xxx
+  #               ref: https://github.com/google/jax/issues/13559#issuecomment-1343573764
+  echo "$0: num_nodes is $num_nodes, proc_per_node is $num_gpus"
   torchrun --nnodes=$num_nodes --nproc_per_node=$num_gpus \
-           --rdzv_id=2023 --rdzv_backend="c10d" --rdzv_endpoint=$HOST_NODE_ADDR \
+           --rdzv_id=$job_id --rdzv_backend="c10d" --rdzv_endpoint=$HOST_NODE_ADDR \
     wenet/bin/train.py \
       --train_engine ${train_engine} \
       --config $train_config \
